updating scripts
1. updated data to 2.0-patch data in ReadAct
2. updated according to changes in 2.0-patch branch in ReadAct (for example, 'name_lang' to 'language', columns' order, and so on)
3. separated all the entity tables' processing into individual modules
4. improved some codes (use regex, use pandas build-in function, etc.)
see #63
