+++ src/scripts/authenticity_space.py
@@ -34,7 +34,7 @@ WHERE {{
 
 
 def read_space_csv(
-    space_url="https://raw.githubusercontent.com/readchina/ReadAct/master/csv/data/Space.csv",
+    space_url="https://raw.githubusercontent.com/readchina/ReadAct/2.0-RC-patch/csv/data/Space.csv",
 ):
     """
     A function to read "Space.csv" for now.
@@ -48,7 +48,7 @@ def read_space_csv(
         if row[0] not in geo_code_dict:
             # key: space_id
             # value: space_name, space_type, lat, lang
-            geo_code_dict[row[0]] = [row[2], row[3], row[5], row[6]]
+            geo_code_dict[row[0]] = [row[3], row[2], row[5], row[6]]
         else:
             print("Space id not exist ?!")
     return geo_code_dict
@@ -72,17 +72,17 @@ def query_with_OSM(k, v):
     if v[0] != "unknown" and v[2] != 0.0:
         lat = str(v[2])
         long = str(v[3])
-    url = (
-        "https://nominatim.openstreetmap.org/reverse?format=xml&lat="
-        + lat
-        + "&lon="
-        + long
-        + "&zoom=18&addressdetails=1&format=json&accept-language=en"
-    )
-    data = requests.get(url)
-    if v[0].lower() not in str(data.json()).lower():
-        item = v + [k]
-        return item
+        url = (
+            "https://nominatim.openstreetmap.org/reverse?format=xml&lat="
+            + lat
+            + "&lon="
+            + long
+            + "&zoom=18&addressdetails=1&format=json&accept-language=en"
+        )
+        data = requests.get(url)
+        if v[0].lower() not in str(data.json()).lower():
+            item = v + [k]
+            return item
 
 
 def geo_code_compare(no_match_list):
@@ -95,7 +95,6 @@ def geo_code_compare(no_match_list):
     space_with_QID = {}  # To collect QIDs for Space.csv
     count = 0
     for i in no_match_list:
-
         if i[0] is None:
             res = None
         else:
@@ -203,54 +202,15 @@ def chunks(it, size):
 
 if __name__ == "__main__":
 
-    # To compare the extracting coordinate location with the info in Space.csv
-    space_url = (
-        "https://raw.githubusercontent.com/readchina/ReadAct/master/csv/data/Space.csv"
-    )
-    geo_code_dict = read_space_csv(space_url)
-
-    # To filter CSV entries with comparing to openstreetmap first
-    no_match_list = compare_to_openstreetmap(geo_code_dict)
-
-    # To compare the rest with wikidata info
-    all_still_no_match_list = []
-    dictionary_list = []
-    for chunk in chunks(no_match_list, 30):  # the digit here controls the batch size
-        if len(chunk) > 0:
-            l, d = geo_code_compare(chunk)
-            if l is not None:
-                all_still_no_match_list += l
-            dictionary_list += [d]
-            print("\n I am taking a break XD \n")
-            time.sleep(
-                10
-            )  # for every a few  entries, let this script take a break of 90 seconds
-    print("Finished the whole iteration")
-    print(all_still_no_match_list)
-    print("dictionary_list", dictionary_list)
-    match_for_space = {k: v for x in dictionary_list for k, v in dict(x).items()}
-    print(match_for_space)
-    with open("../results/match_for_space.json", "w") as f:
-        json.dump(match_for_space, f)
-
-    # """
-    # The following is a one-time script to deactivate the compare_to_openstreetmap function to get as much Q-ids as
-    # we can.
-    # """
-    #
     # # To compare the extracting coordinate location with the info in Space.csv
     # space_url = (
-    #     "https://raw.githubusercontent.com/readchina/ReadAct/master/csv/data/Space.csv"
+    #     "https://raw.githubusercontent.com/readchina/ReadAct/2.0-RC-patch/csv/data/Space.csv"
     # )
     # geo_code_dict = read_space_csv(space_url)
     #
-    # # Assume 0 match from openstreetmap
-    # no_match_list = []
-    # for k, v in geo_code_dict.items():
-    #     if v[0] != "unknown" and v[2] != 0.0:
-    #         item = v + [k]
-    #         no_match_list.append(item)
-    #
+    # # To filter CSV entries with comparing to openstreetmap first
+    # no_match_list = compare_to_openstreetmap(geo_code_dict)
+
     # # To compare the rest with wikidata info
     # all_still_no_match_list = []
     # dictionary_list = []
@@ -269,5 +229,44 @@ if __name__ == "__main__":
     # print("dictionary_list", dictionary_list)
     # match_for_space = {k: v for x in dictionary_list for k, v in dict(x).items()}
     # print(match_for_space)
-    # with open("../results/match_for_space_without_Openstreetmap.json", "w") as f:
+    # with open("../results/match_for_space.json", "w") as f:
     #     json.dump(match_for_space, f)
+
+    """
+    The following is a one-time script to deactivate the compare_to_openstreetmap function to get as much Q-ids as
+    we can.
+    """
+
+    # To compare the extracting coordinate location with the info in Space.csv
+    space_url = "https://raw.githubusercontent.com/readchina/ReadAct/2.0-RC-patch/csv/data/Space.csv"
+    geo_code_dict = read_space_csv(space_url)
+
+    # Assume 0 match from openstreetmap
+    no_match_list = []
+    for k, v in geo_code_dict.items():
+        if v[0] != "unknown" and v[2] != 0.0:
+            item = v + [k]
+            no_match_list.append(item)
+
+    # To compare the rest with wikidata info
+    all_still_no_match_list = []
+    dictionary_list = []
+    for chunk in chunks(no_match_list, 30):  # the digit here controls the batch size
+        if len(chunk) > 0:
+            l, d = geo_code_compare(chunk)
+            if l is not None:
+                all_still_no_match_list += l
+            dictionary_list += [d]
+
+            print("\n I am taking a break XD \n")
+            time.sleep(
+                10
+            )  # for every a few  entries, let this script take a break of 90 seconds
+
+    print("Finished the whole iteration")
+    print("all_still_no_match_list:", all_still_no_match_list)
+    print("dictionary_list", dictionary_list)
+    match_for_space = {k: v for x in dictionary_list for k, v in dict(x).items()}
+    print(match_for_space)
+    with open("../results/match_for_space_without_Openstreetmap.json", "w") as f:
+        json.dump(match_for_space, f)
